Idea of implementation
We miss somemthing essential for the gossipElector to work, we need to have the same pubKey for the validator and the peer (which is not the case)
For this implementation to work we have to find a way to link both the reactor with the peer of the instance.
Once done we will have the relation between the stake and the peer allowing the send to better nodes.